requirements.txt
Flask==1.1.2
requests==2.24.0
beautifulsoup4==4.9.1
Flask-SQLAlchemy==2.4.4
Flask-Script==2.0.6
Flask-Migrate==2.5.3
psycopg2==2.8.5
celery==4.4.7
redis==3.5.3
python-dotenv==0.14.0